% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_tree.R
\docType{data}
\name{fish_tree}
\alias{fish_tree}
\title{Phylogenetic tree of marine fish hosts of Caligus and Lepeophtheirus}
\format{
A \code{phylo} object with 625 tips.
}
\source{
\url{https://fishtreeoflife.org/}. The underlying data are
  released under the CC0 1.0 public domain dedication (Dryad,
  \doi{10.5061/dryad.fc71cp4}). Please cite Rabosky, D. L. and
  colleagues (2018)
  An inverse latitudinal gradient in speciation rate for marine fishes.
  Nature 559, 392-395. \doi{10.1038/s41586-018-0273-1}
}
\usage{
fish_tree
}
\description{
A phylogenetic tree (class \code{phylo}) of 625 marine fish species that
are recorded hosts of sea lice of the genera \emph{Caligus} and
\emph{Lepeophtheirus} in the \code{cofid} copepod-fish interaction
database. The tree is pruned from the Fish Tree of Life
actinopterygian phylogeny (Rabosky and colleagues 2018,
actinopt_12k_raxml).
Tip labels are binomial species names with spaces.
}
\seealso{
The \code{cofid} package for the interaction records.
}
\keyword{datasets}
