% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_dist.R
\docType{data}
\name{phy_dist}
\alias{phy_dist}
\title{Phylogenetic distance matrix among host data on phylopred}
\format{
A matrix with 235 rows and 235 columns.
Contain Phylogenetic distance matrix between bark beetle host genera
}
\source{
\url{https://www.barkbeetles.info/}
}
\usage{
phy_dist
}
\description{
Data of phylogenetic distance matrix for bark beetles host species
}
\keyword{datasets}
